Crate dense_mats

This crate contains traits to abstract over dense matrices.

The goal is to enable interoperability between various linear algebra libraries, by providing a common definition.

Matrices are thus defined as a contiguous piece of memory, the data array, which is interpreted as a matrix stored in either column major order or row major order.

Enums

StorageOrder
Describe the storage order of a matrix.

Traits

DenseMatOwned
Represents a dense matrix we own

DenseMatView
Represents a view into a dense matrix

DenseMatViewMut
A mutable view into a dense matrix