Enum dense_mats::StorageOrder

pub enum StorageOrder {
    F,
    C,
    Unordered,
}

Describe the storage order of a matrix.

Variants

F

C storage order, ie column major storage for matrices The dimensions are sorted in decreasing order of variation

C

C storage order, ie row major storage for matrices The dimensions are sorted in increasing order of variation

Unordered

Nothing special can be assumed about the order