Module dense_mats::tensor

Structs

Axis
A type for indexing an axis of a tensor

ChunkOuterBlocks
An iterator over non-overlapping blocks of a tensor, along the least-varying dimension

ChunkOuterBlocksMut
An iterator over non-overlapping mutable blocks of a matrix, along the least-varying dimension

Slices

SlicesMut

Tensor
A simple dense matrix

Type Definitions

MatOwned

MatView

MatViewMut

TensorOwned

TensorView

TensorViewMut

VecOwned

VecView

VecViewMut