Trait dense_mats::DenseMatView

pub trait DenseMatView<N> {
    fn rows(&self) -> usize;
    fn cols(&self) -> usize;
    fn strides(&self) -> [usize; 2];
    fn data(&self) -> &[N];

    fn shape(&self) -> [usize; 2] { ... }
    fn ordering(&self) -> StorageOrder { ... }
    fn data_index(&self, i: usize, j: usize) -> usize { ... }
}

Represents a view into a dense matrix

Required Methods

fn rows(&self) -> usize

The number of rows of the matrix

fn cols(&self) -> usize

The number of cols of the matrix

fn strides(&self) -> [usize; 2]

The strides of the matrix.

self.strides()[0] gives the number of elements that must be skipped into self.data() to get to the element of the next row with the same column. self.strides()[1] gives the number of elements that must be skipped into self.data() to get to the element of the next column with the same row.

For a row major matrix of shape (3, 4) with contiguous storage, the strides would be [4, 1].

For alignement reasons, it is possible to have strides that don't match the shape of the matrix (meaning that some elements of the data array are unused).

fn data(&self) -> &[N]

Access to the matrix's data

Getting access to the element located at row i and column j can be done by indexing self.data() at the location computed by i * strides[0] + j * strides[1]

Provided Methods

fn shape(&self) -> [usize; 2]

The number of rows and cols of the matrix

fn ordering(&self) -> StorageOrder

Storage order. Specifies which dimension is stored the most contiguously in memory

fn data_index(&self, i: usize, j: usize) -> usize

Give the index into self.data() for accessing the element at row i and column j