Crate dense_mats

This crate contains strided dense tensor implementations.

Tensors are defined as a contiguous piece of memory, the data array, which is interpreted using its stride information.

Reexports

pub use tensor::{Tensor, MatView, MatViewMut, MatOwned};

pub use tensor::{VecView, VecViewMut, VecOwned};

Modules

array_like
This module is a simple workaround around the fact that arrays don't implement Copy or Deref (even though in practice they are). This enables asking for arrays as a generic bound

errors
Error type for sprs

tensor

Enums

StorageOrder
Describe the storage order of a matrix.