Crate dense_mats

This crate contains strided dense matrices implementations.

Matrices are defined as a contiguous piece of memory, the data array, which is interpreted as a matrix stored in either column major order or row major order.

Modules

errors
Error type for sprs

Structs

StridedMat
A simple dense matrix

StridedVec
A simple dense vector

Enums

StorageOrder
Describe the storage order of a matrix.

Type Definitions

StridedMatOwned

StridedMatView

StridedMatViewMut

StridedVecOwned

StridedVecView

StridedVecViewMut